DOCUMENTATION + DIFFERENT KEYBINDS
